mqtt_listener/requirements.txt
paho-mqtt==1.6.1
pycryptodome==3.19.0
requests